+++ unbrake-frontend/src/Configuration/ConfigFunctions.jsx
@@ -1,0 +1,19 @@
+export const createConfig = data => {
+  const configurationDefault = {
+    CONFIG_ENSAIO: {
+      LSL: data.inferiorLimit,
+      LWT: data.inferiorTime,
+      NOS: data.number,
+      TAO: data.enableOutput,
+      TAS: data.temperature,
+      TAT: data.time,
+      TBS: data.timeBetweenCycles,
+      TMO: data.disableShutdown,
+      USL: data.upperLimit,
+      UWT: data.upperTime
+    }
+  };
+  return configurationDefault;
+};
+
+export default createConfig;
